♻️ Realizado a divisão de responsabilidades da classe Main para uma classe de gerenciamento de tasks específica.
